requirements/dev.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
coverage==5.3
Django==3.1.2
django-redis==4.12.1
djangorestframework==3.12.1
importlib-metadata==2.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.2
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.3.1
